refactor(baseball): Divide class
